refactor: move MethodDescriptor out of JvmName
